+++ lib/glimmer/swt/image_proxy.rb
@@ -122,9 +122,11 @@ module Glimmer
       def input_stream
         if @file_path.start_with?('uri:classloader')
           @jar_file_path = @file_path
-          file_path = @jar_file_path.sub(/^uri\:classloader\:/, '').sub('//', '') # the latter sub is needed for Mac
-          project_name = java.lang.System.getProperty('project_name')
-          file_input_stream = java_import("#{project_name}.Resource").last.java_class.resource_as_stream(file_path)
+          file_path = @jar_file_path.sub(/^uri\:classloader\:/, '').sub(/^\/+/, '')
+          require 'jruby'
+          jcl = JRuby.runtime.jruby_class_loader
+          resource = jcl.get_resource_as_stream(file_path)
+          file_input_stream = resource.to_io.to_input_stream
         else
           file_input_stream = java.io.FileInputStream.new(@file_path)
         end